02._data_formats/python/translationServers/translationServers/pyproject.toml
[tool.poetry]
name = "translationservers"
version = "0.1.0"
description = ""
authors = ["Raszbanana <75260673+Raszbanana@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
xmltodict = "^0.13.0"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
